+++ src/api/service.tsx
@@ -5,6 +5,7 @@ import {
     IEditServiceInfo,
     IServiceEditTime,
     IServiceShowHide,
+    IUpdateService,
 } from "../interfaces/services/Iservice";
 import useSWR from "swr";
 import { app_api } from "../helper/url";
@@ -30,7 +31,7 @@ export const addService = async (serviceData: InsertService, token: string) => {
 
 export const updateService = async (
     serviceId: number,
-    serviceData: any,
+    serviceData: IUpdateService,
     token: string
 ) => {
     token = token.replace(/"/g, "");
@@ -58,8 +59,7 @@ export const getServiceByBusinessId = async (
     token = token.replace(/"/g, "");
     try {
         const services = await axios.get(
-            `${
-                import.meta.env.VITE_APP_API
+            `${import.meta.env.VITE_APP_API
             }/getListServiceByBusinessId/${businessId}?page=1&limit=10`,
             {
                 headers: {
@@ -79,8 +79,7 @@ export const deleteService = async (serviceId: number, token: string) => {
     token = token.replace(/"/g, "");
     try {
         const response = await axios.put(
-            `${
-                import.meta.env.VITE_APP_API
+            `${import.meta.env.VITE_APP_API
             }/deleteServiceByHidden/${serviceId}`,
             {},
             {
@@ -91,30 +90,6 @@ export const deleteService = async (serviceId: number, token: string) => {
         );
 
         return response.data;
-    } catch (error) {
-        console.error(error);
-        throw error;
-    }
-};
-
-export const getServiceByServiceId = async (
-    serviceId: number,
-    token: string
-) => {
-    token = token.replace(/"/g, "");
-    try {
-        const services = await axios.get(
-            `${
-                import.meta.env.VITE_APP_API
-            }/getServiceByServiceId/${serviceId}`,
-            {
-                headers: {
-                    Authorization: token,
-                },
-            }
-        );
-
-        return services.data as any;
     } catch (error) {
         console.error(error);
         throw error;
@@ -175,8 +150,7 @@ export const updateServiceShowHide = async (
     token = token.replace(/"/g, "");
     try {
         const business = await axios.post(
-            `${
-                import.meta.env.VITE_APP_API
+            `${import.meta.env.VITE_APP_API
             }/updateServiceShowHide/${serviceId}`,
             serviceData,
             {
@@ -193,6 +167,7 @@ export const updateServiceShowHide = async (
 };
 
 export const getServices = () => {
+    // eslint-disable-next-line react-hooks/rules-of-hooks
     const { data, error, isLoading } = useSWR(
         `${app_api}/services`,
         (url: string) =>
