--- src/pages/business/components/OpenMap.tsx
@@ -1,5 +1,0 @@
-// import React from "react";
-
-export default function OpenMap() {
-    return <div>OpenMap</div>;
-}
