+++ src/pages/service/BusinessPreview.tsx
@@ -33,7 +33,7 @@ interface BusinessPreviewProps {
     bookingSlots: {
         daysOpen: string[];
         availableFromDate: string;
-        availableToDate: string;
+        availableToDate: string | null;
         slotsTime: {
             startTime: string;
             endTime: string;
@@ -49,6 +49,8 @@ interface BusinessPreviewProps {
 
 const BusinessPreview = (props: BusinessPreviewProps) => {
     const { t } = useTranslation();
+
+    console.log("props", props);
 
     const [calendar] = useState({
         // handle calendar date
@@ -132,11 +134,11 @@ const BusinessPreview = (props: BusinessPreviewProps) => {
                         <div
                             // key={index}
                             className={`border-2 rounded-lg p-5 border-[#003B95] bg-[#006CE31A] text-[#003B95]`}
-                            // ${
-                            //   item.isSelected
-                            //     ? "border-[#003B95] bg-[#006CE31A] text-[#003B95]"
-                            //     : ""
-                            // }
+                        // ${
+                        //   item.isSelected
+                        //     ? "border-[#003B95] bg-[#006CE31A] text-[#003B95]"
+                        //     : ""
+                        // }
                         >
                             <div className="">
                                 <p className="flex justify-between">
@@ -223,14 +225,13 @@ const BusinessPreview = (props: BusinessPreviewProps) => {
                                 <div
                                     key={index}
                                     className={`h-[72px] flex flex-col justify-center items-center border rounded-lg cursor-pointer
-                  ${
-                      moment().isSame(item, "day")
-                          ? "border-2 border-[#003B95] bg-[#006CE31A] text-[#003B95]"
-                          : ""
-                  }`}
-                                    // ${
-                                    //   isOpen ? "" : "text-[#8B8B8B] bg-[#8B8B8B] bg-opacity-20"
-                                    // }
+                  ${moment().isSame(item, "day")
+                                            ? "border-2 border-[#003B95] bg-[#006CE31A] text-[#003B95]"
+                                            : ""
+                                        }`}
+                                // ${
+                                //   isOpen ? "" : "text-[#8B8B8B] bg-[#8B8B8B] bg-opacity-20"
+                                // }
                                 >
                                     <p className="text-[14px] font-thin">
                                         {item.format("dd")}
@@ -252,46 +253,42 @@ const BusinessPreview = (props: BusinessPreviewProps) => {
                         <p>{t("time")}</p>
                     </h2>
                     <div
-                        className={`grid gap-4 mt-2 ${
-                            slotArrays
-                                ? slotArrays?.slotsTime.length === 1
-                                    ? "grid-cols-1"
-                                    : "grid-cols-2"
-                                : ""
-                        }`}
+                        className={`grid gap-4 mt-2 ${slotArrays
+                            ? slotArrays?.slotsTime.length === 1
+                                ? "grid-cols-1"
+                                : "grid-cols-2"
+                            : ""
+                            }`}
                     >
                         {slotArrays && slotArrays.slotsTime.length > 0 ? (
                             slotArrays?.slotsTime.map((ii: any, jj: number) => {
                                 return (
                                     <div
                                         key={jj}
-                                        className={`flex flex-col border-2 rounded-lg text-center p-3 ${
-                                            ii?.capacity >= 1
-                                                ? " cursor-pointer"
-                                                : "text-[#8C8C8C] bg-[#8B8B8B33]"
-                                        } ${
-                                            ii.isSelected
+                                        className={`flex flex-col border-2 rounded-lg text-center p-3 ${ii?.capacity >= 1
+                                            ? " cursor-pointer"
+                                            : "text-[#8C8C8C] bg-[#8B8B8B33]"
+                                            } ${ii.isSelected
                                                 ? "bg-[#006CE31A] border-[#003B95] text-[#003B95]"
                                                 : ""
-                                        }`}
+                                            }`}
                                     >
                                         <span>
                                             {ii.startTime} - {ii.endTime}
                                         </span>
                                         <span className="flex justify-center items-center text-[12px]">
                                             <div
-                                                className={`h-[9px] w-[9px] mx-1 rounded ${
-                                                    ii?.capacity < 5
-                                                        ? ii?.capacity === 0
-                                                            ? "bg-[#9C9C9C]"
-                                                            : "bg-[#FEC84B]"
-                                                        : "bg-[#12B76A]"
-                                                }`}
+                                                className={`h-[9px] w-[9px] mx-1 rounded ${ii?.capacity < 5
+                                                    ? ii?.capacity === 0
+                                                        ? "bg-[#9C9C9C]"
+                                                        : "bg-[#FEC84B]"
+                                                    : "bg-[#12B76A]"
+                                                    }`}
                                             ></div>
                                             {ii?.capacity !== 0
                                                 ? `${ii?.capacity} ${t(
-                                                      "available"
-                                                  )}`
+                                                    "available"
+                                                )}`
                                                 : t("full")}
                                         </span>
                                     </div>
