+++ src/pages/service/ServiceDetail.tsx
@@ -1,8 +1,6 @@
-import { alpha, Drawer } from "@mui/material";
-import AddCircleOutlineIcon from "@mui/icons-material/AddCircleOutline";
 import { useContext, useEffect, useState } from "react";
-import { updateService } from "../../api/service";
-import { Divider } from "@mui/material";
+import { addService, updateService } from "../../api/service";
+import { Divider, alpha } from "@mui/material";
 import { useTranslation } from "react-i18next";
 import Header from "./components/Header";
 import SettingServiceBtn from "./components/SettingServiceBtn";
@@ -14,17 +12,20 @@ import AddServiceTime from "./AddServiceTime";
 import useSWR from "swr";
 import { app_api, fetcher } from "../../helper/url";
 import Loading from "../../components/dialog/Loading";
-import { useParams } from "react-router-dom";
+import { useNavigate, useParams } from "react-router-dom";
 import ConfirmCard from "../../components/dialog/ConfirmCard";
-import { Anchor } from "./ServiceSetting";
 import BusinessPreview from "./BusinessPreview";
 import {
     IServiceInfo,
     IServiceEditTime,
+    IServiceByServiceId,
 } from "../../interfaces/services/Iservice";
 import { GlobalContext } from "../../contexts/BusinessContext";
 import toast from "react-hot-toast";
 import CheckCircleOutlineIcon from "@mui/icons-material/CheckCircleOutline";
+import AddServiceInfo from "./AddServiceInfo";
+import { IBusinessesById } from "../../interfaces/business";
+import AddCircleOutlineIcon from "@mui/icons-material/AddCircleOutline";
 
 interface IParams {
     serviceId: number;
@@ -35,7 +36,10 @@ interface IParams {
 export default function ServiceDetail(props: IParams) {
     const token = localStorage.getItem("token");
     const { businessId } = useParams();
+    const navigate = useNavigate();
     const { t } = useTranslation();
+    const queryParams = new URLSearchParams(location.search);
+    const type = queryParams.get("type");
 
     const { setShowDialog, setDialogState } = useContext(GlobalContext);
 
@@ -47,44 +51,45 @@ export default function ServiceDetail(props: IParams) {
     const [isEditTime, setIsEditTime] = useState(false);
     const [selectedIndex, setSelectedIndex] = useState(0);
     const [isAddTime, setIsAddTime] = useState(false);
+    const [isOpenPreview, setIsOpenPreview] = useState(false);
+    const [isCreateService, setIsCreateService] = useState(true);
+    const [isCloseCard, setIsCloseCard] = useState(false);
     const [modifyServiceInfo, setModifyServiceInfo] = useState<IServiceInfo>(); // for update service info
     const [modifyServiceTime, setModifyServiceTime] =
-        useState<IServiceEditTime[]>(); // for update service time
+        useState<IServiceEditTime[]>([]); // for update service time
     const [openConfirm, setOpenConfirm] = useState(false);
     const {
         data: serviceInfo,
         isLoading: serviceLoading,
         mutate: serviceMutate,
-    } = useSWR<any>(
+    } = useSWR<IServiceByServiceId>(
         `${app_api}/getServiceByServiceId/${props.serviceId}`,
         fetcher
     );
-    const [state, setState] = useState({
-        top: false,
-        left: false,
-        bottom: false,
-        right: false,
-    });
-    // const [isSuccess, setIsSuccess] = useState(false);
+    const { data: businessData } = useSWR<IBusinessesById>(
+        businessId && `${app_api}/business/${businessId}`,
+        fetcher
+    );
 
     useEffect(
         () => {
-            if (serviceInfo) {
-                setModifyServiceInfo({
-                    serviceName: serviceInfo.title,
-                    serviceDescription: serviceInfo.description,
-                    price: serviceInfo.price,
-                    currency: serviceInfo.currency,
-                });
-                setModifyServiceTime(serviceInfo.bookingSlots);
-                setIsHideEndTime(serviceInfo.isHideEndTime);
-                setIsAutoApprove(serviceInfo.isAutoApprove);
-                setIsHidePrice(serviceInfo.isHidePrice);
+            if (!type) {
+                if (serviceInfo) {
+                    setModifyServiceInfo({
+                        serviceName: serviceInfo.title,
+                        serviceDescription: serviceInfo.description,
+                        price: serviceInfo.price,
+                        currency: serviceInfo.currency,
+                    });
+                    setModifyServiceTime(serviceInfo.bookingSlots);
+                    setIsHideEndTime(serviceInfo.isHideEndTime);
+                    setIsAutoApprove(serviceInfo.isAutoApprove);
+                    setIsHidePrice(serviceInfo.isHidePrice);
+                }
             }
         },
         // eslint-disable-next-line react-hooks/exhaustive-deps
-        [serviceLoading]
-    );
+        [serviceLoading]);
 
     const handleSetEditInfo = () => {
         setIsEditInfo(!isEditInfo);
@@ -95,37 +100,28 @@ export default function ServiceDetail(props: IParams) {
     const handleAddTime = () => {
         setIsAddTime(!isAddTime);
     };
-
-    const handleUpdateService = async () => {
-        const insertData = {
-            title: modifyServiceInfo?.serviceName,
-            description: modifyServiceInfo?.serviceDescription,
-            price: modifyServiceInfo?.price,
-            currency: modifyServiceInfo?.currency,
-            bookingSlots: modifyServiceTime,
-            isAutoApprove: isAutoApprove,
-            isHidePrice: isHidePrice,
-            isHideEndTime: isHideEndTime,
-            businessId: Number(businessId),
-        };
-
-        await updateService(serviceInfo.id, insertData, token || "");
-        setShowDialog(false);
-        toast(t("editSuccess"), {
-            icon: <CheckCircleOutlineIcon sx={{ color: "green" }} />,
-        });
-        setDialogState("phone-input");
-        props.serviceMutate && props.serviceMutate();
-        serviceMutate();
-        props.handleClose && props.handleClose();
+    const handleOpenPreview = () => {
+        setIsOpenPreview(!isOpenPreview);
+    };
+    const handleOpenServiceInfo = () => {
+        setIsCreateService(!isCreateService);
     };
 
     const handleDeleteServiceTime = async () => {
-        const updatedTimeDetails = serviceInfo.bookingSlots.filter(
-            (_item: IServiceEditTime, i: number) => i !== selectedIndex
-        );
-        setModifyServiceTime(updatedTimeDetails);
-        serviceMutate();
+        if (type) {
+            const updatedTimeDetails = modifyServiceTime.filter(
+                (_item: IServiceEditTime, i: number) => i !== selectedIndex
+            );
+            setModifyServiceTime(updatedTimeDetails);
+        } else {
+            if (serviceInfo) {
+                const updatedTimeDetails = serviceInfo.bookingSlots.filter(
+                    (_item: IServiceEditTime, i: number) => i !== selectedIndex
+                );
+                setModifyServiceTime(updatedTimeDetails);
+                serviceMutate();
+            }
+        }
     };
 
     const handleSetServiceInfo = (serviceInFo: IServiceInfo) => {
@@ -135,42 +131,31 @@ export default function ServiceDetail(props: IParams) {
         setModifyServiceTime(serviceTime);
     };
 
-    const toggleDrawer =
-        (anchor: Anchor, open: boolean) =>
-        (event: React.KeyboardEvent | React.MouseEvent) => {
-            if (
-                event.type === "keydown" &&
-                ((event as React.KeyboardEvent).key === "Tab" ||
-                    (event as React.KeyboardEvent).key === "Shift")
-            ) {
-                return;
-            }
-            setState({ ...state, [anchor]: open });
-        };
-
     const handleIsModifiedData = () => {
-        return (
-            JSON.stringify({
-                title: serviceInfo.title,
-                description: serviceInfo.description,
-                price: serviceInfo.price,
-                currency: serviceInfo.currency,
-                bookingSlots: serviceInfo.bookingSlots,
-                isAutoApprove: serviceInfo.isAutoApprove,
-                isHidePrice: serviceInfo.isHidePrice,
-                isHideEndTime: serviceInfo.isHideEndTime,
-            }) !==
-            JSON.stringify({
-                title: modifyServiceInfo?.serviceName,
-                description: modifyServiceInfo?.serviceDescription,
-                price: modifyServiceInfo?.price,
-                currency: modifyServiceInfo?.currency,
-                bookingSlots: modifyServiceTime,
-                isAutoApprove: isAutoApprove,
-                isHidePrice: isHidePrice,
-                isHideEndTime: isHideEndTime,
-            })
-        );
+        if (serviceInfo) {
+            return (
+                JSON.stringify({
+                    title: serviceInfo.title,
+                    description: serviceInfo.description,
+                    price: serviceInfo.price,
+                    currency: serviceInfo.currency,
+                    bookingSlots: serviceInfo.bookingSlots,
+                    isAutoApprove: serviceInfo.isAutoApprove,
+                    isHidePrice: serviceInfo.isHidePrice,
+                    isHideEndTime: serviceInfo.isHideEndTime,
+                }) !==
+                JSON.stringify({
+                    title: modifyServiceInfo?.serviceName,
+                    description: modifyServiceInfo?.serviceDescription,
+                    price: modifyServiceInfo?.price,
+                    currency: modifyServiceInfo?.currency,
+                    bookingSlots: modifyServiceTime,
+                    isAutoApprove: isAutoApprove,
+                    isHidePrice: isHidePrice,
+                    isHideEndTime: isHideEndTime,
+                })
+            );
+        }
     };
 
     const handleCloseCardDetail = () => {
@@ -186,31 +171,82 @@ export default function ServiceDetail(props: IParams) {
         props.handleClose && props.handleClose();
     };
 
-    const previewService = () => (
-        <BusinessPreview
-            businessId={Number(businessId)}
-            title={modifyServiceInfo?.serviceName ?? ""}
-            description={modifyServiceInfo?.serviceDescription ?? ""}
-            price={modifyServiceInfo?.price ?? 0}
-            isAutoApprove={isAutoApprove}
-            currency={modifyServiceInfo?.currency ?? ""}
-            bookingSlots={modifyServiceTime ?? []}
-            availableFromDate={""}
-            availableToDate={""}
-            isHidePrice={isHidePrice}
-            isHideEndTime={isHideEndTime}
-            handleClose={() => setState({ ...state, bottom: false })}
-        />
-    );
+    const ismodifyServiceInfoUndefined = () => {
+        return (
+            modifyServiceInfo?.serviceName === undefined ||
+            modifyServiceInfo?.serviceDescription === undefined ||
+            modifyServiceInfo?.price === undefined ||
+            modifyServiceInfo?.currency === undefined
+        );
+    };
 
+    const handleCreateUpdateService = async () => {
+        if (type && modifyServiceInfo && modifyServiceTime && token && businessData) {
+            const insertData = {
+                businessId: Number(businessId ?? ""),
+                title: modifyServiceInfo.serviceName,
+                description: modifyServiceInfo.serviceDescription,
+                price: modifyServiceInfo.price,
+                isAutoApprove: isAutoApprove,
+                currency: modifyServiceInfo.currency,
+                openTime: businessData.openTime.substring(0, 5),
+                closeTime: businessData.closeTime.substring(0, 5),
+                bookingSlots: modifyServiceTime.map((time) => ({
+                    daysOpen: time.daysOpen,
+                    availableFromDate: time.availableFromDate,
+                    availableToDate:
+                        time.availableToDate === "" ? null : time.availableToDate,
+                    slotsTime: time.slotsTime,
+                    duration: time.duration,
+                })),
+                availableFromDate: modifyServiceTime[0].availableFromDate,
+                availableToDate:
+                    modifyServiceTime[0].availableToDate === ""
+                        ? null
+                        : modifyServiceTime[0].availableToDate,
+                isHidePrice: isHidePrice,
+                isHideEndTime: isHideEndTime,
+            };
+            await addService(insertData, token).then(() => {
+                props.serviceMutate && props.serviceMutate();
+                localStorage.removeItem("serviceInfo");
+                localStorage.removeItem("serviceTime");
+                toast(t("addService"), {
+                    icon: <CheckCircleOutlineIcon sx={{ color: "green" }} />,
+                });
+                if (type && type == "create") {
+                    navigate(`/business-profile/${businessId}`);
+                } else {
+                    navigate(`/service-setting/${businessId}`);
+                }
+            });
+        } else {
+            if (serviceInfo && modifyServiceInfo && modifyServiceTime && token && businessData) {
+                const insertData = {
+                    title: modifyServiceInfo?.serviceName,
+                    description: modifyServiceInfo?.serviceDescription,
+                    price: modifyServiceInfo?.price,
+                    currency: modifyServiceInfo?.currency,
+                    bookingSlots: modifyServiceTime,
+                    isAutoApprove: isAutoApprove,
+                    isHidePrice: isHidePrice,
+                    isHideEndTime: isHideEndTime,
+                    businessId: Number(businessId),
+                };
+                await updateService(serviceInfo.id, insertData, token);
+                setShowDialog(false);
+                toast(t("editSuccess"), {
+                    icon: <CheckCircleOutlineIcon sx={{ color: "green" }} />,
+                });
+                setDialogState("phone-input");
+                props.serviceMutate && props.serviceMutate();
+                serviceMutate();
+                props.handleClose && props.handleClose();
+            }
+        }
+    };
     return (
         <>
-            <Drawer
-                anchor={"bottom"}
-                open={state["bottom"]}
-                onClose={toggleDrawer("bottom", false)}>
-                {previewService()}
-            </Drawer>
             <ConfirmCard
                 open={openConfirm}
                 handleClose={() => setOpenConfirm(false)}
@@ -220,34 +256,49 @@ export default function ServiceDetail(props: IParams) {
                 bntConfirm={t("button:discard")}
                 bntBack={t("button:back")}
             />
-            <Loading openLoading={serviceLoading} />
-            {modifyServiceInfo &&
-                (isEditInfo ? (
-                    <EditServiceInfo
-                        serviceName={modifyServiceInfo.serviceName}
-                        serviceDescription={
-                            modifyServiceInfo.serviceDescription
-                        }
-                        price={modifyServiceInfo.price}
-                        currency={modifyServiceInfo.currency}
-                        handleSetEditInfo={handleSetEditInfo}
-                        serviceMutate={serviceMutate}
-                        handleSetServiceInfo={handleSetServiceInfo}
+            {type ? (
+                isAddTime ? (
+                    <AddServiceTime
+                        isAddServiceTime={isAddTime}
+                        serviceTime={modifyServiceTime ?? []}
+                        handleAddTime={handleAddTime}
+                        handleCloseCreateService={handleOpenServiceInfo}
                     />
+                ) : isCreateService ? (
+                    <AddServiceInfo
+                        serviceInfo={modifyServiceInfo}
+                        isClose={isCloseCard}
+                        isEdit={isEditInfo}
+                        handleAddTime={handleAddTime}
+                        handleSetServiceInfo={handleSetServiceInfo}
+                        handleClose={handleOpenServiceInfo}
+                    />
+                ) : isOpenPreview ? (
+                    <>
+                        <BusinessPreview
+                            businessId={Number(businessId)}
+                            title={modifyServiceInfo?.serviceName ?? ""}
+                            description={modifyServiceInfo?.serviceDescription ?? ""}
+                            price={modifyServiceInfo?.price ?? 0}
+                            isAutoApprove={isAutoApprove}
+                            currency={modifyServiceInfo?.currency ?? ""}
+                            bookingSlots={modifyServiceTime ?? []}
+                            availableFromDate={""}
+                            availableToDate={""}
+                            isHidePrice={isHidePrice}
+                            isHideEndTime={isHideEndTime}
+                            handleClose={handleOpenPreview}
+                        />
+                    </>
                 ) : isEditTime ? (
                     <EditServiceTime
                         serviceTime={modifyServiceTime ?? []}
-                        openTime={serviceInfo.openTime}
-                        closeTime={serviceInfo.closeTime}
+                        openTime={businessData?.openTime ?? ""}
+                        closeTime={businessData?.closeTime ?? ""}
                         editIndex={selectedIndex}
                         isAddTime={isAddTime}
                         handleSetEditTime={handleSetEditTime}
                         handleSetServiceTime={handleSetServiceTime}
-                    />
-                ) : isAddTime ? (
-                    <AddServiceTime
-                        serviceTime={serviceInfo.bookingSlots}
-                        handleAddTime={handleAddTime}
                     />
                 ) : (
                     <div
@@ -255,24 +306,28 @@ export default function ServiceDetail(props: IParams) {
                         style={{ width: "100vw" }}>
                         <div className="pr-4 pl-4 pt-6">
                             <Header
+                                isClose={false}
                                 context={t("title:serviceInformation")}
-                                handleClose={handleCloseCardDetail}
+                                handleClose={handleSetEditTime}
                             />
                         </div>
                         <Divider sx={{ marginTop: "16px", width: "100%" }} />
                         <div className="flex flex-col pr-4 pl-4 mb-[80px]">
                             <div className="mt-4 flex flex-col gap-3">
-                                <ServiceCard
-                                    serviceId={serviceInfo.id}
-                                    serviceName={modifyServiceInfo.serviceName}
-                                    serviceDescription={
-                                        modifyServiceInfo.serviceDescription
-                                    }
-                                    price={modifyServiceInfo.price}
-                                    currency={modifyServiceInfo.currency}
-                                    handleSetEditInfo={handleSetEditInfo}
-                                />
-
+                                {modifyServiceInfo &&
+                                    <ServiceCard
+                                        serviceName={modifyServiceInfo.serviceName}
+                                        serviceDescription={
+                                            modifyServiceInfo.serviceDescription
+                                        }
+                                        price={modifyServiceInfo.price}
+                                        currency={modifyServiceInfo.currency}
+                                        handleSetEditInfo={() => {
+                                            handleOpenServiceInfo();
+                                            setIsEditInfo(true);
+                                            setIsCloseCard(true);
+                                        }}
+                                    />}
                                 {modifyServiceTime &&
                                     modifyServiceTime.map(
                                         (
@@ -286,7 +341,7 @@ export default function ServiceDetail(props: IParams) {
                                                         item.availableFromDate
                                                     }
                                                     availableToDate={
-                                                        item.availableToDate
+                                                        item.availableToDate ?? ""
                                                     }
                                                     slotsTime={item.slotsTime}
                                                     selectedIndex={index}
@@ -316,24 +371,20 @@ export default function ServiceDetail(props: IParams) {
                                         borderRadius: "8px",
                                         justifyContent: "center",
                                     }}
+                                    onClick={handleAddTime}
                                     className=" items-center gap-1 p-1 ">
                                     <AddCircleOutlineIcon
                                         sx={{ fontSize: "13px" }}
                                     />
                                     <div
-                                        className=" font-medium "
-                                        onClick={handleAddTime}>
+                                        className=" font-medium ">
                                         {t("button:addServiceTime")}
                                     </div>
                                 </button>
 
-                                <p
-                                    className=" font-bold "
-                                    style={{ fontSize: "14px" }}>
-                                    {t("serviceSetting")}
-                                </p>
-
                                 <SettingServiceBtn
+                                    modifyServiceInfo={modifyServiceInfo}
+                                    modufyServiceTime={modifyServiceTime}
                                     isAutoApprove={isAutoApprove}
                                     isHidePrice={isHidePrice}
                                     isHideEndTime={isHideEndTime}
@@ -346,26 +397,166 @@ export default function ServiceDetail(props: IParams) {
                                     handleHideEndTime={() =>
                                         setIsHideEndTime(!isHideEndTime)
                                     }
+                                    handleUpdateService={handleCreateUpdateService}
+                                    handleOpenPreview={handleOpenPreview}
                                 />
-
-                                <div
-                                    className={`w-full flex justify-center bottom-0 inset-x-0 gap-2 mb-3 fixed`}>
-                                    <button
-                                        className="w-[45vw] p-3 border text-deep-blue border-deep-blue rounded-lg font-semibold"
-                                        onClick={toggleDrawer("bottom", true)}>
-                                        {t("button:preview")}
-                                    </button>
-                                    <button
-                                        onClick={handleUpdateService}
-                                        type="submit"
-                                        className="w-[45vw] p-3 text-white bg-deep-blue rounded-lg font-semibold">
-                                        {t("button:confirm")}
-                                    </button>
-                                </div>
                             </div>
                         </div>
                     </div>
-                ))}
+                )
+            ) : (
+                <>
+                    <Loading openLoading={serviceLoading} />
+                    {(!ismodifyServiceInfoUndefined() && modifyServiceInfo && serviceInfo) && (
+                        (isEditInfo ? (
+                            <EditServiceInfo
+                                serviceName={modifyServiceInfo.serviceName}
+                                serviceDescription={
+                                    modifyServiceInfo.serviceDescription
+                                }
+                                price={modifyServiceInfo.price}
+                                currency={modifyServiceInfo.currency}
+                                handleSetEditInfo={handleSetEditInfo}
+                                serviceMutate={serviceMutate}
+                                handleSetServiceInfo={handleSetServiceInfo}
+                            />
+                        ) : isEditTime ? (
+                            <EditServiceTime
+                                serviceTime={modifyServiceTime ?? []}
+                                openTime={serviceInfo.openTime}
+                                closeTime={serviceInfo.closeTime}
+                                editIndex={selectedIndex}
+                                isAddTime={isAddTime}
+                                handleSetEditTime={handleSetEditTime}
+                                handleSetServiceTime={handleSetServiceTime}
+                            />
+                        ) : isAddTime ? (
+                            <AddServiceTime
+                                isAddServiceTime={false}
+                                serviceTime={serviceInfo.bookingSlots}
+                                handleAddTime={handleAddTime}
+                            />
+                        ) : isOpenPreview ? (
+                            <>
+                                <BusinessPreview
+                                    businessId={Number(businessId)}
+                                    title={modifyServiceInfo?.serviceName ?? ""}
+                                    description={modifyServiceInfo?.serviceDescription ?? ""}
+                                    price={modifyServiceInfo?.price ?? 0}
+                                    isAutoApprove={isAutoApprove}
+                                    currency={modifyServiceInfo?.currency ?? ""}
+                                    bookingSlots={modifyServiceTime ?? []}
+                                    availableFromDate={""}
+                                    availableToDate={""}
+                                    isHidePrice={isHidePrice}
+                                    isHideEndTime={isHideEndTime}
+                                    handleClose={handleOpenPreview}
+                                />
+                            </>
+                        ) : (
+                            <div
+                                className={`w-full sm:w-auto md:w-full lg:w-auto xl:w-full overflow-x-hidden`}
+                                style={{ width: "100vw" }}>
+                                <div className="pr-4 pl-4 pt-6">
+                                    <Header
+                                        isClose={false}
+                                        context={t("title:serviceInformation")}
+                                        handleClose={handleCloseCardDetail}
+                                    />
+                                </div>
+                                <Divider sx={{ marginTop: "16px", width: "100%" }} />
+                                <div className="flex flex-col pr-4 pl-4 mb-[80px]">
+                                    <div className="mt-4 flex flex-col gap-3">
+                                        <ServiceCard
+                                            serviceId={serviceInfo.id}
+                                            serviceName={modifyServiceInfo.serviceName}
+                                            serviceDescription={
+                                                modifyServiceInfo.serviceDescription
+                                            }
+                                            price={modifyServiceInfo.price}
+                                            currency={modifyServiceInfo.currency}
+                                            handleSetEditInfo={handleSetEditInfo}
+                                        />
+
+                                        {modifyServiceTime &&
+                                            modifyServiceTime.map(
+                                                (
+                                                    item: IServiceEditTime,
+                                                    index: number
+                                                ) => (
+                                                    <div key={index}>
+                                                        <TimeCard
+                                                            daysOpen={item.daysOpen}
+                                                            availableFromDate={
+                                                                item.availableFromDate
+                                                            }
+                                                            availableToDate={
+                                                                item.availableToDate ?? ""
+                                                            }
+                                                            slotsTime={item.slotsTime}
+                                                            selectedIndex={index}
+                                                            handleSetEditTime={
+                                                                handleSetEditTime
+                                                            }
+                                                            handleSelectIndex={(
+                                                                index: number
+                                                            ) =>
+                                                                setSelectedIndex(index)
+                                                            }
+                                                            handleDeleteServiceTime={
+                                                                handleDeleteServiceTime
+                                                            }
+                                                        />
+                                                    </div>
+                                                )
+                                            )}
+
+                                        <button
+                                            style={{
+                                                display: "flex",
+                                                background: `${alpha("#020873", 0.1)}`,
+                                                width: "135px",
+                                                height: "27px",
+                                                fontSize: "14px",
+                                                borderRadius: "8px",
+                                                justifyContent: "center",
+                                            }}
+                                            onClick={handleAddTime}
+                                            className=" items-center gap-1 p-1 ">
+                                            <AddCircleOutlineIcon
+                                                sx={{ fontSize: "13px" }}
+                                            />
+                                            <div
+                                                className=" font-medium ">
+                                                {t("button:addServiceTime")}
+                                            </div>
+                                        </button>
+
+                                        <SettingServiceBtn
+                                            modifyServiceInfo={modifyServiceInfo}
+                                            modufyServiceTime={modifyServiceTime}
+                                            isAutoApprove={isAutoApprove}
+                                            isHidePrice={isHidePrice}
+                                            isHideEndTime={isHideEndTime}
+                                            handleAutoApprove={() =>
+                                                setIsAutoApprove(!isAutoApprove)
+                                            }
+                                            handleHidePrice={() =>
+                                                setIsHidePrice(!isHidePrice)
+                                            }
+                                            handleHideEndTime={() =>
+                                                setIsHideEndTime(!isHideEndTime)
+                                            }
+                                            handleUpdateService={handleCreateUpdateService}
+                                            handleOpenPreview={handleOpenPreview}
+                                        />
+
+                                    </div>
+                                </div>
+                            </div>
+                        )))}
+                </>
+            )}
         </>
     );
 }
